// src/app/shared/services/policy/entities/specialCondition.dto.ts
/**
* conditions  DTO.ts
*
* @description: condttions of policies
* @author Andres Tamayo
* @date 11-02-2019.
*
**/
export interface SpecialConditionOutputDto {
    limitationName: string;
    effectiveDate: string;
    description: string;
    languageId: number;
​​​​    memberId: number;
}